Add a "gendoc" test program and add testing infrastructure
The purpose of this new "gendoc" program is to take Ada source code,
turn it into a Prettier Document, and then use the Prettier engine to
turn it to reformatted Ada source code.

The (for now only) test driver runs gendoc and checks its Document and
reformatted Ada output against baselines.

At this stage, "gendoc" is just stubbed: its outputs are dummy ones, and
so the only testcase fails. Future work will implement the actual
Ada-to-Document and Document-to-reformatted-Ada conversions (see the
"TODO???" comments left in gendoc.adb.