Remove XFAIL
The reason neither foo.ads nor bar.ads are formatted is because they are not
withed by any source of the default.gpr project.

This is intended and consistent with what libgpr2 does (we're using its API for
project loading and switches parsing). If the user wants to really format all
files, even the ones not used, then he should pass the -U flag.